# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Mini social network with users, public/private posts, and friend requests",
        "title": "SocialNet API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/friendships/friends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "List accepted friends",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.FriendsListResponse"
                        }
                    }
                }
            }
        },
        "/v1/friendships/requests": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "Send a friend request",
                "parameters": [
                    {
                        "description": "Friend request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.FriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.FriendshipResponse"
                        }
                    }
                }
            }
        },
        "/v1/friendships/requests/received": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "List pending received friend requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.FriendshipsListResponse"
                        }
                    }
                }
            }
        },
        "/v1/friendships/requests/sent": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "List pending sent friend requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.FriendshipsListResponse"
                        }
                    }
                }
            }
        },
        "/v1/friendships/requests/{id}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "Accept a friend request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.FriendshipResponse"
                        }
                    }
                }
            }
        },
        "/v1/friendships/requests/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "Reject a friend request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.FriendshipResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_donca_user-crud_internal_friendships_models.FriendProfile": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_donca_user-crud_internal_friendships_models.FriendRequest": {
            "type": "object",
            "properties": {
                "addressee_id": {
                    "type": "string"
                }
            }
        },
        "github_com_donca_user-crud_internal_friendships_models.FriendsListResponse": {
            "type": "object",
            "properties": {
                "friends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.FriendProfile"
                    }
                }
            }
        },
        "github_com_donca_user-crud_internal_friendships_models.Friendship": {
            "type": "object",
            "properties": {
                "addressee_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "requester_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_donca_user-crud_internal_friendships_models.FriendshipResponse": {
            "type": "object",
            "properties": {
                "friendship": {
                    "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.Friendship"
                }
            }
        },
        "github_com_donca_user-crud_internal_friendships_models.FriendshipsListResponse": {
            "type": "object",
            "properties": {
                "friendships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_donca_user-crud_internal_friendships_models.Friendship"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token. Format: Bearer \u0026lt;token\u0026gt;",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}